Split UI elements into individual files
